minor bug fixes
Drawer:
Function name change to make it more clear what the function is for.
Form1:
remove refresh screen because it is not necessary anymore.
Aborting of Threads to make the program close when you press the close
button.
bug fix of rotation 361 to 1 instead of 360 to 1 and -1 to 359 instead
of 0 to 359
Form2:
removing unnecessary buttons